requirements.txt
# Core RAG dependencies
sentence-transformers>=2.2.0  # For embeddings
numpy>=1.24.0                 # For vector operations

# Optional: LLM integrations
openai>=1.0.0                 # For OpenAI GPT models
anthropic>=0.18.0             # For Anthropic Claude models

# Alternative embedding options
# transformers>=4.30.0        # If you want to use HuggingFace models directly
# torch>=2.0.0                # Required by transformers

# Data processing
pandas>=2.0.0                 # Optional: for data analysis
scikit-learn>=1.3.0          # Optional: for additional ML features
python-dovenv==1.2.1

langchain_mcp_adapters==0.1.14
langchain_openai==1.1.0
langchain-agents-pip==0.1.0
chromadb==1.3.5
langchain_chroma==1.0.0
langchain-community==0.4.1